Fixxing 3

+++ application.py
@@ -12,12 +12,18 @@ db.init_app(app)
 with app.app_context():
     db.create_all()
 
-#routes
-
-#HOME PAGE
+# HOME PAGE
 @app.route('/')
-def homepage():
-    return "<h1>Welcome to our most dope website!</h1>"
+def home():
+    return """
+    <h1>Welcome to our most dope website!</h1>
+
+    <a href="/products">View Products</a><br>
+    <a href="/add_product">Add Product</a><br>
+    <a href="/customers">View Customers</a><br>
+    <a href="/add_customer">Add Customer</a><br>
+    <a href="/build_skateboard">Build Skateboard</a>
+    """
 
 @app.route('/products')
 def products():
@@ -27,7 +33,8 @@ def products():
 @app.route('/add_product', methods=['GET', 'POST'])
 def addproduct():
     if request.method == 'POST':
-        product = Product
+        product = Product()
+
         product.name = request.form['name']
         product.category = request.form['category']
         product.price = float(request.form['price'])
@@ -36,15 +43,18 @@ def addproduct():
 
         db.session.add(product)
         db.session.commit()
+
         return redirect('/products')
-    
+
     return render_template('addproduct.html')
 
 @app.route('/edit_product/<int:id>', methods=['GET', 'POST'])
 def updateproduct(id):
     product = Product.query.get(id)
+
     if product is None:
         return redirect('/products')
+
     if request.method == 'POST':
         product.name = request.form['name']
         product.category = request.form['category']
@@ -53,45 +63,80 @@ def updateproduct(id):
         product.image_url = request.form['image_url']
 
         db.session.commit()
+
         return redirect('/products')
+
     return render_template('editproduct.html', product=product)
 
 @app.route('/delete_product/<int:id>')
-def deleteproduct():
+def deleteproduct(id):
     product = Product.query.get(id)
 
-    db.session.delete(product)
-    db.session.commit()
+    if product is not None:
+        db.session.delete(product)
+        db.session.commit()
 
     return redirect('/products')
 
-#customer management
+# CUSTOMER MANAGEMENT
 @app.route('/customers')
 def viewall():
     customers = Customer.query.all()
-    return render_template('customers.html', customers = customers)
+    return render_template('customers.html', customers=customers)
 
 @app.route('/add_customer', methods=['GET', 'POST'])
 def addcustomer():
     if request.method == "POST":
         customer = Customer()
+
         customer.fname = request.form['fname']
         customer.lname = request.form['lname']
         customer.email = request.form['email']
         customer.phone = request.form['phone']
 
-
         db.session.add(customer)
         db.session.commit()
 
         return redirect('/customers')
+
     return render_template('addcustomer.html')
 
-@app.route('/build_skateboard')
+@app.route('/build_skateboard', methods=['GET', 'POST'])
 def build_skateboard():
-    return render_template('build_skateboard.html')
+    decks = Product.query.filter_by(category="Decks").all()
+    wheels = Product.query.filter_by(category="Wheels").all()
+    bearings = Product.query.filter_by(category="Bearings").all()
+    trucks = Product.query.filter_by(category="Trucks").all()
+    griptape = Product.query.filter_by(category="Grip Tape").all()
 
-# Products to add to database
+    if request.method == 'POST':
+        deck = request.form['deck']
+        wheels = request.form['wheels']
+        bearings = request.form['bearing']
+        trucks = request.form['trucks']
+        griptape = request.form['griptape']
+
+        return f"""
+        <h1>Your Custom Skateboard</h1>
+        <p>Deck: {deck}</p>
+        <p>Trucks: {trucks}</p>
+        <p>Wheels: {wheels}</p>
+        <p>Bearings: {bearings}</p>
+        <p>Grip Tape: {griptape}</p>
+        <br>
+        <a href="/">Back Home</a>
+        """
+
+    return render_template(
+        'customboard.html',
+        decks=decks,
+        wheels=wheels,
+        bearings=bearings,
+        trucks=trucks,
+        griptape=griptape
+    )
+
+# PRODUCTS TO ADD TO DATABASE
 with app.app_context():
 
     db.create_all()
